+++ pytorch_lightning/metrics/functional/__init__.py
@@ -11,6 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+# TODO: unify metrics between class and functional, add below
+from pytorch_lightning.metrics.functional.accuracy import accuracy  # noqa: F401
 from pytorch_lightning.metrics.functional.average_precision import average_precision  # noqa: F401
 from pytorch_lightning.metrics.functional.classification import (  # noqa: F401
     auc,
@@ -18,18 +20,13 @@ from pytorch_lightning.metrics.functional.classification import (  # noqa: F401
     dice_score,
     get_num_classes,
     multiclass_auroc,
-    precision,
-    precision_recall,
-    recall,
     stat_scores_multiple_classes,
     to_categorical,
     to_onehot,
 )
-# TODO: unify metrics between class and functional, add below
-from pytorch_lightning.metrics.functional.accuracy import accuracy  # noqa: F401
 from pytorch_lightning.metrics.functional.confusion_matrix import confusion_matrix  # noqa: F401
 from pytorch_lightning.metrics.functional.explained_variance import explained_variance  # noqa: F401
-from pytorch_lightning.metrics.functional.f_beta import fbeta, f1  # noqa: F401
+from pytorch_lightning.metrics.functional.f_beta import f1, fbeta  # noqa: F401
 from pytorch_lightning.metrics.functional.hamming_distance import hamming_distance  # noqa: F401
 from pytorch_lightning.metrics.functional.ir_average_precision import retrieval_average_precision  # noqa: F401
 from pytorch_lightning.metrics.functional.image_gradients import image_gradients  # noqa: F401
@@ -38,6 +35,7 @@ from pytorch_lightning.metrics.functional.mean_absolute_error import mean_absolu
 from pytorch_lightning.metrics.functional.mean_squared_error import mean_squared_error  # noqa: F401
 from pytorch_lightning.metrics.functional.mean_squared_log_error import mean_squared_log_error  # noqa: F401
 from pytorch_lightning.metrics.functional.nlp import bleu_score  # noqa: F401
+from pytorch_lightning.metrics.functional.precision_recall import precision, precision_recall, recall  # noqa: F401
 from pytorch_lightning.metrics.functional.precision_recall_curve import precision_recall_curve  # noqa: F401
 from pytorch_lightning.metrics.functional.psnr import psnr  # noqa: F401
 from pytorch_lightning.metrics.functional.r2score import r2score  # noqa: F401