+++ tests/core/test_memory.py
@@ -16,7 +16,7 @@ import torch
 import torch.nn as nn
 
 from pytorch_lightning import LightningModule, Trainer
-from pytorch_lightning.core.memory import UNKNOWN_SIZE, ModelSummary
+from pytorch_lightning.core.memory import ModelSummary, UNKNOWN_SIZE
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base.models import ParityModuleRNN